+++ Cron/Feed.php
@@ -7,7 +7,7 @@ namespace Salesfire\Salesfire\Cron;
  *
  * @category   Salesfire
  * @package    Salesfire_Salesfire
- * @version.   1.2.3
+ * @version.   1.2.4
  */
 class Feed
 {
@@ -61,6 +61,10 @@ class Feed
     public function execute()
     {
         $storeCollection = $this->_storeManager->getStores();
+
+        $object_manager = \Magento\Framework\App\ObjectManager::getInstance();
+        $stock_state = $object_manager->get('\Magento\CatalogInventory\Api\StockStateInterface');
+
         foreach ($storeCollection as $store)
         {
             $storeId = $store->getId();
@@ -84,7 +88,7 @@ class Feed
             $currency = $store->getCurrentCurrencyCode();
 
             $this->printLine($siteId, '<?xml version="1.0" encoding="utf-8" ?>', 0);
-            $this->printLine($siteId, '<productfeed site="'.$this->_storeManager->getStore($storeId)->getBaseUrl(\Magento\Framework\UrlInterface::URL_TYPE_WEB).'" date-generated="'.gmdate('c').'">', 0);
+            $this->printLine($siteId, '<productfeed site="'.$this->_storeManager->getStore($storeId)->getBaseUrl(\Magento\Framework\UrlInterface::URL_TYPE_WEB).'" date-generated="'.gmdate('c').'" version="' . $this->_helperData->getVersion() . '">', 0);
 
             $mediaUrl = $this->_storeManager->getStore($storeId)->getBaseUrl(\Magento\Framework\UrlInterface::URL_TYPE_MEDIA);
 
@@ -247,19 +251,21 @@ class Feed
 
                                 $this->printLine($siteId, '<mpn><![CDATA['.$this->escapeString($childProduct->getSku()).']]></mpn>', 5);
 
-                                $stock_item = $this->_stockItem->get($childProduct->getId());
-                                $this->printLine($siteId, '<stock>'.($stock_item && $stock_item->getIsInStock() ? ($stock_item->getQty() > 0 ? (int) $stock_item->getQty() : 1) : 0).'</stock>', 5);
+                                $stock = null;
+                                try {
+                                    $stock_item = $this->_stockItem->get($childProduct->getId());
+                                    $stock = ($stock_item && $stock_item->getIsInStock()) ? ($stock_item->getQty() > 0 ? (int) $stock_item->getQty() : 1) : 0;
+                                } catch (\Magento\Framework\Exception\NoSuchEntityException $e) {
+                                    $stock = $stock_state->getStockQty($childProduct->getId());
+                                }
+
+                                $this->printLine($siteId, '<stock>' . ($stock ? $stock : 0) .'</stock>', 5);
 
                                 $this->printLine($siteId, '<link>' . $product->getProductUrl(true) . '</link>', 5);
 
-                                $image = $childProduct->getImage();
-                                if (! empty($image) && $image != 'no_selection') {
-                                    $this->printLine($siteId, '<image>' . $mediaUrl . 'catalog/product' . $image . '</image>', 5);
-                                } else {
-                                    $image = $product->getImage();
-                                    if (! empty($image) && $image != 'no_selection') {
-                                        $this->printLine($siteId, '<image>' . $mediaUrl . 'catalog/product' . $image . '</image>', 5);
-                                    }
+                                $image = $this->getProductImage($siteId, $mediaUrl, $product, $childProduct);
+                                if (! empty($image)) {
+                                    $this->printLine($siteId, '<image>' . $image  . '</image>', 5);
                                 }
 
                                 $this->printLine($siteId, '</variant>', 4);
@@ -304,14 +310,21 @@ class Feed
 
                         $this->printLine($siteId, '<mpn><![CDATA['.$this->escapeString($product->getSku()).']]></mpn>', 5);
 
-                        $stock_item = $this->_stockItem->get($product->getId());
-                        $this->printLine($siteId, '<stock>'.($stock_item && $stock_item->getIsInStock() ? ($stock_item->getMinQty() > 0 ? (int) $stock_item->getQty() : 1) : 0).'</stock>', 5);
+                        $stock = null;
+                        try {
+                            $stock_item = $this->_stockItem->get($product->getId());
+                            $stock = ($stock_item && $stock_item->getIsInStock()) ? ($stock_item->getQty() > 0 ? (int) $stock_item->getQty() : 1) : 0;
+                        } catch (\Magento\Framework\Exception\NoSuchEntityException $e) {
+                            $stock = $stock_state->getStockQty($product->getId());
+                        }
+
+                        $this->printLine($siteId, '<stock>' . ($stock ? $stock : 0) .'</stock>', 5);
 
                         $this->printLine($siteId, '<link>' . $product->getProductUrl(true) . '</link>', 5);
 
-                        $image = $product->getImage();
+                        $image = $this->getProductImage($siteId, $mediaUrl, $product, $product);
                         if (! empty($image)) {
-                            $this->printLine($siteId, '<image>' . $mediaUrl . 'catalog/product' . $image . '</image>', 5);
+                            $this->printLine($siteId, '<image>' . $image  . '</image>', 5);
                         }
 
                         $this->printLine($siteId, '</variant>', 4);
@@ -391,6 +404,8 @@ class Feed
 
         $collection->clear();
 
+        $collection->addMediaGalleryData();
+
         return $collection;
     }
 
@@ -437,11 +452,30 @@ class Feed
         if(! empty($attribute_obj)) {
             $attribute_text = $attribute_obj->setStoreId($storeId)->getFrontend()->getValue($product);
 
-            if (! empty($attribute_text)) {
+            if (! empty($attribute_text) && $attribute_text != 'no_selection') {
                 return $attribute_text;
             }
         }
 
         return null;
+    }
+
+    protected function getProductImage($siteId, $mediaUrl, $product, $childProduct) {
+        $image = $childProduct->getImage();
+        if (empty($image) || $image == 'no_selection') {
+            $image = $product->getImage();
+        }
+
+        if (empty($image) || $image == 'no_selection') {
+            $image = $product->getMediaGalleryImages()->getFirstItem()->getUrl();
+        } else {
+            $image = $mediaUrl . 'catalog/product' . $image;
+        }
+
+        if (empty($image) || $image == 'no_selection') {
+            return null;
+        }
+
+        return $image;
     }
 }