+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### Salesfire v1.2.4
+Released on 2020-02-10
+Release notes:
+
+- Add product image fallback
+- Add verison to feed
+- Fixed stock level
+
+
 ### Salesfire v1.2.3
 Released on 2020-02-07
 Release notes: